requirements.txt
networkx==3.1
rank-bm25==0.2.2
scikit-learn>=1.4.0
numpy>=1.26.0
pypdf==4.2.0
openai==1.30.1
python-dotenv==1.0.1
requests==2.31.0
pdfplumber==0.11.0
langchain==0.2.0
langchain-community==0.2.0
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
bibtexparser>=1.4.0
rispy>=0.9.0
click>=8.0.0

